Configure の内部クラスを export
